requirements.txt
Django~=3.1.6
djangorestframework
django-cors-headers
mysqlclient